remove logic from component
