+++ md_editor/components/EditorStats.js
@@ -1,6 +1,9 @@
 import React, { useState, useEffect } from "react";
 import DocCounterCard from "./DocCounterCard";
 
+import docStats from "../logic/mdDocWordCount";
+import docTagCounter from "../logic/mdDocTagCount";
+
 function TagsInDocument(props) {
   const tagValues = Object.keys(props.htmlTagGroup);
 
@@ -28,77 +31,11 @@ function TagsInDocument(props) {
 }
 
 export default function EditorActions(props) {
+  // what if we put the clean list up here
   const [documentRowCount, updateDocRowCount] = useState(0);
   const [documentWordCount, updateDocWordCount] = useState(0);
   const [documentLetterCount, updateLetterCount] = useState(0);
   const [documentTags, updateDocTags] = useState({});
-
-  function htmlTags(mdString) {
-    const htmlTag = /<(\w+)/.exec(mdString);
-    return htmlTag[1];
-  }
-
-  function wordsInRow(mdString) {
-    // regex returning an array
-    // index 0 will be the full html string
-    // index 1 will be the group with out tag'
-    const stingGroup = />(.*)</.exec(mdString);
-    const wordCount = stingGroup[1].trim().split(" ");
-    return wordCount;
-  }
-
-  function reduceNumberCount(list) {
-    const sumOfNumber = list.reduce((total, current) => {
-      return total + current;
-    });
-    return sumOfNumber;
-  }
-
-  function wordsPerDoc(mdInput) {
-    let numWordsPerRow = mdInput.map((x) => {
-      return x.length;
-    });
-    const totalNumbWordCount = reduceNumberCount(numWordsPerRow);
-    return totalNumbWordCount;
-  }
-
-  function letterPerDoc(mdInput) {
-    const letterPerRow = mdInput.map((x) => x.join("").length);
-    const letterTotal = reduceNumberCount(letterPerRow);
-    return letterTotal;
-  }
-
-  function docStats(mdStrings) {
-    const htmlInputString = mdStrings.map((x) => {
-      // array of words,
-      return wordsInRow(x);
-    });
-
-    // Number of Word in Document
-    const numbOfDocWords = wordsPerDoc(htmlInputString);
-
-    //Number of Letter in Document
-    const numbOfLetters = letterPerDoc(htmlInputString);
-
-    const GOAL = {
-      wordCount: numbOfDocWords,
-      letterCount: numbOfLetters,
-    };
-    return GOAL;
-  }
-
-  function tagOBject(mdHtmlTags) {
-    // make the tag
-    const tagsAmount = {};
-    for (let i = 1; i < mdHtmlTags.length; i++) {
-      if (mdHtmlTags[i] in tagsAmount) {
-        tagsAmount[mdHtmlTags[i]] += 1;
-      } else {
-        tagsAmount[mdHtmlTags[i]] = 1;
-      }
-    }
-    return tagsAmount;
-  }
 
   useEffect(() => {
     if (props.FullMkList.length != 0 && props.FullMkList[0] != undefined) {
@@ -108,11 +45,8 @@ export default function EditorActions(props) {
       const replaceNewLine = removeEmptyIndex.map((x) =>
         x.replace(/(\r\n|\n|\r)/gm, " ")
       );
-      const htmlTagslist = replaceNewLine.map((htmlString) => {
-        return htmlTags(htmlString);
-      });
 
-      const tagCounter = tagOBject(htmlTagslist);
+      const tagCounter = docTagCounter(replaceNewLine);
       const docCounts = docStats(replaceNewLine);
 
       updateDocTags(tagCounter);
